fixup! refactor the whole structure
