--- src/aead/gcm.rs
@@ -1,7 +1,0 @@
-pub struct Aes128Gcm;
-pub struct Aes256Gcm;
-
-#[cfg(feature = "tls12")]
-pub mod tls12;
-
-pub mod tls13;
