+++ src/lib.rs
@@ -81,28 +81,24 @@ pub const ALL_CIPHER_SUITES: &[SupportedCipherSuite] = misc::const_concat_slices
     {
         #[cfg(feature = "tls12")]
         {
-            TLS12_SUITES
+            tls12::suites::TLS12_SUITES
         }
         #[cfg(not(feature = "tls12"))]
         {
             &[]
         }
     },
-    TLS13_SUITES,
+    tls13::suites::TLS13_SUITES,
 );
 
-mod aead;
-mod hash;
-mod hmac;
-mod kx;
-mod misc;
+pub mod aead;
+pub mod hash;
+pub mod hmac;
+pub mod kx;
+pub mod misc;
 pub mod sign;
-mod verify;
+pub mod verify;
 
 #[cfg(feature = "tls12")]
 pub mod tls12;
 pub mod tls13;
-
-#[cfg(feature = "tls12")]
-pub use tls12::*;
-pub use tls13::*;
