+++ src/sign.rs
@@ -2,8 +2,8 @@
 use alloc::{sync::Arc, vec::Vec};
 use core::marker::PhantomData;
 
-use self::ecdsa::{EcdsaSigningKeyP256, EcdsaSigningKeyP384};
-use self::eddsa::Ed25519SigningKey;
+use self::ecdsa::nist::{EcdsaSigningKeyP256, EcdsaSigningKeyP384};
+use self::eddsa::ed25519::Ed25519SigningKey;
 use self::rsa::RsaSigningKey;
 
 use pki_types::PrivateKeyDer;
