+++ src/verify/eddsa.rs
@@ -1,2 +1,1 @@
-mod ed25519;
-pub use ed25519::*;
+pub mod ed25519;
